requirements.txt
tkinter
PyQt5
sqlite3
cryptography
pyinstaller
pillow
reportlab
datetime
os
shutil
json
hashlib
platform
winreg
uuid
pyinstaller==6.1.0
Pillow==10.1.0
reportlab==4.0.7